Move the subscribed_events fixture to be shared
[skip ci]
